feat: add quiz question count

+++ src/main/java/com/juan/itb/vrbackend/repository/ResponseRepository.java
@@ -12,5 +12,5 @@ public interface ResponseRepository extends R2dbcRepository<Response, Long> {
   @Query("SELECT COUNT(*) FROM responses r " +
       "JOIN options o ON r.option_id = o.id " +
       "WHERE r.user_id = :userId AND r.quiz_id = :quizId AND o.is_correct = TRUE")
-  Mono<Integer> calculateScore(Long userId, Long quizId);
+  Mono<Long> calculateScore(Long userId, Long quizId);
 }
